refactor: :recycle: removal of "Glob" booleans through addition of Glob Enum.
The Glob enum has been introduced in order to get rid of duplicated settings. This makes the Checking of the provided arguments easier in source code since Enums are Exhaustive. This comes at a minimal usage friendliness degradation since now it would be needed to specify "--glob insensitive" instead of "--iglob". Even then I think it is worth it :3