Fix poor performance when using --long
Previously, the unix extended attributes are always read no matter
what. This is not ideal, as it is very expensive to call for files
which we will not display and leads to significantly (~80x with level
= 2) worse runtimes recursing large directories. This is fixed by only
calling `unix::Attrs::from` when it is a file which is going to be
displayed.